requirements.txt
pygame==1.9.6
parl==1.3.1
paddlepaddle>=1.8.4
mock==1.0.1
numpydoc
Sphinx==1.2.3
sphinx_rtd_theme